requirements.txt
numpy
pandas
scikit-learn
pyyaml
nltk
lightgbm
awscli
dvc[s3]
mlflow
dagshub
python-dotenv
matplotlib
seaborn
